gestion_fichier/.idea/requirements.txt
# pip install -r requirements.txt
Flask==2.2.3
Flask_JWT_Extended==4.5.2
flask_sqlalchemy==3.0.3
Werkzeug==2.2.2